add support for openharmony
With these fixes, mozjs compiles successfully for
the target 'aarch64-linux-ohos'. This also requires
using a custom build of rustc with support for this
target, and clang 10 from the ohos-sdk native toolchain.

Additional wrappers are needed for the clang wrappers to
be invoked with the sysroot from the SDK as well.

Wrappers:
https://doc.rust-lang.org/nightly/rustc/platform-support/openharmony.html

ohos-sdk with clang toolchain:
https://gitee.com/openharmony/docs/blob/master/en/release-notes/OpenHarmony-v3.2.2-release.md

prebuilt rustc:
clone https://gitee.com/openharmony/build.git
and run the prebuilts_download.sh